+++ team_first_project01/.eslintrc.json
@@ -19,4 +19,3 @@
       "no-unused-vars": ["warn", { "vars": "all", "args": "after-used", "ignoreRestSiblings": false }]
     }
   }
-
