+++ team_first_project01/views/board.ejs
@@ -6,7 +6,7 @@
 </head>
 <body>
 
-<form action="/upload/create" method="post" enctype="multipart/form-data">
+<form action="/productimagefile" method="get" enctype="multipart/form-data">
     <input type="file" name="imgFile">
     <input type="submit" value="파일 업로드하기">
 </form>
